requirements.txt
experta==1.9.4
frozendict==1.2
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
schema==0.6.7
six==1.16.0
tzdata==2023.3
